Exceptionok gyártássának folytatása, implementálása a megfelelő helyen.

+++ src/main/java/hu/progmasters/finalexam/quidditch/exceptionhandling/ClubNotFoundException.java
@@ -2,13 +2,13 @@ package hu.progmasters.finalexam.quidditch.exceptionhandling;
 
 public class ClubNotFoundException extends RuntimeException {
 
-    private final int CLUB_ID;
+    private final long CLUB_ID;
 
-    public ClubNotFoundException(int CLUB_ID) {
+    public ClubNotFoundException(long CLUB_ID) {
         this.CLUB_ID = CLUB_ID;
     }
 
-    public int getCLUB_ID() {
+    public long getCLUB_ID() {
         return CLUB_ID;
     }
 }
